Added a new function that calculates the Metropolis-Hastings acceptance ratio where the Boltzmann Factors are all calculated in log-space. If the log of the values is between log(1/RAND_MAX) and 0, the value is exponentiated. Else we have 0 or 2 (greater than 1).

+++ src/energy.c
@@ -38,9 +38,9 @@ float Energy_InitPotential(int beadID) {
                     tmpR[j] = tmpR[j] - nBoxSize[j] / 2;
                     totEn += (float) (tmpR[j] * tmpR[j]);
                 }
-                if (totEn >= 250) {
-                    totEn = (fCuTemp - fKT) * ((float) totEn);
-                } else if (totEn <= 600) {
+                if (totEn >= 250.) {
+                    totEn = (fCuTemp - fKT) * ( totEn);
+                } else if (totEn <= 600.) {
                     totEn = (fCuTemp - fKT) * ((float) tot_beads + 1. / (totEn + 0.02));
                 }
                 else{
